Tests/LibWeb/Text/expected/css/fit-content-interpolation.txt

fit-content(14px)
fit-content(16px)
